migrate hl7 json lake to deprecated
